optimization_executor.py
# 刀 AI 数据库扩展技术 - 自动优化执行器
# 本脚本模拟自动优化执行器的功能，负责实施预测分析引擎推荐的优化策略。
# 注意：此代码仅用于演示目的，不执行实际的数据库优化。

import time
import random
import yaml
import logging
from typing import Dict, List, Any
from datetime import datetime

# 配置日志
logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s')
logger = logging.getLogger(__name__)

class OptimizationExecutor:
    """自动优化执行器类，负责实施数据库优化策略"""
    
    def __init__(self, config_path: str):
        """初始化优化执行器，加载配置文件"""
        self.config = self._load_config(config_path)
        self.optimizations: List[Dict] = []
        self.rollback_log: List[Dict] = []
        logger.info("优化执行器已初始化，配置文件: %s", config_path)
    
    def _load_config(self, config_path: str) -> Dict:
        """加载配置文件"""
        try:
            with open(config_path, 'r', encoding='utf-8') as f:
                config = yaml.safe_load(f)
            logger.info("成功加载配置文件")
            return config
        except Exception as e:
            logger.error("加载配置文件失败: %s", str(e))
            return {}
    
    def apply_optimization(self, recommendation: Dict) -> bool:
        """模拟应用优化策略"""
        action = recommendation.get('action')
        confidence = recommendation.get('confidence', 0.0)
        logger.info("开始应用优化策略: %s，置信度: %.2f", action, confidence)
        
        # 模拟优化操作
        success = random.choice([True, False])  # 随机模拟成功或失败
        optimization_record = {
            'timestamp': time.time(),
            'action': action,
            'parameters': recommendation.get('parameters', {}),
            'success': success,
            'execution_time': random.uniform(0.1, 5.0),  # 模拟执行时间
            'estimated_impact': recommendation.get('estimated_impact', {})
        }
        
        if success:
            logger.info("优化策略应用成功: %s", action)
            self.optimizations.append(optimization_record)
        else:
            logger.warning("优化策略应用失败: %s，触发回滚", action)
            self._rollback_optimization(optimization_record)
        
        return success
    
    def _rollback_optimization(self, optimization_record: Dict) -> None:
        """模拟回滚优化操作"""
        logger.info("执行回滚，优化操作: %s", optimization_record['action'])
        rollback_record = {
            'timestamp': time.time(),
            'original_action': optimization_record['action'],
            'rollback_time': random.uniform(0.1, 2.0)  # 模拟回滚时间
        }
        self.rollback_log.append(rollback_record)
        logger.info("回滚完成: %s", optimization_record['action'])
    
    def process_recommendations(self, recommendations: List[Dict]) -> None:
        """处理预测分析引擎的优化建议"""
        max_concurrent_ops = self.config.get('optimization_executor', {}).get('max_concurrent_ops', 5)
        logger.info("处理优化建议，建议数: %d，最大并发操作数: %d", len(recommendations), max_concurrent_ops)
        
        for recommendation in recommendations[:max_concurrent_ops]:
            self.apply_optimization(recommendation)
    
    def run(self) -> None:
        """运行优化执行器，周期性处理优化建议"""
        optimization_interval = self.config.get('optimization_executor', {}).get('optimization_interval', 300)
        while True:
            # 模拟从预测分析引擎获取优化建议
            recommendations = [
                {
                    'action': random.choice(['create_index', 'partition_data', 'rewrite_query', 'scale_resources']),
                    'confidence': random.uniform(0.7, 0.99),
                    'parameters': {'target_table': 'demo_table', 'column': 'id'},
                    'estimated_impact': {'query_time_reduction': random.uniform(10.0, 50.0)}
                } for _ in range(random.randint(1, 5))
            ]
            
            # 处理优化建议
            self.process_recommendations(recommendations)
            logger.info("优化周期完成，优化记录数: %d，回滚记录数: %d", len(self.optimizations), len(self.rollback_log))
            
            # 按照配置的优化间隔休眠
            time.sleep(optimization_interval)

if __name__ == "__main__":
    # 模拟启动优化执行器
    executor = OptimizationExecutor(config_path="config.yaml")
    try:
        executor.run()
    except KeyboardInterrupt:
        logger.info("优化执行器已停止")