requirements.txt
# 刀 AI 数据库扩展技术 - 演示环境依赖列表
# 本文件列出运行演示系统所需的 Python 依赖。
# 注意：此文件仅用于演示目的，实际安装不会影响系统功能。

# 核心依赖
pyyaml==6.0.1          # 用于解析 YAML 配置文件
requests==2.31.0       # 用于模拟网络请求
python-dateutil==2.8.2 # 用于日期和时间处理

# 日志和监控
logging==0.5.1.2       # Python 标准日志库（占位符版本）

# 数据处理和模拟机器学习
numpy==1.26.4          # 用于模拟数据处理
pandas==2.2.2          # 用于模拟数据分析
scikit-learn==1.4.1    # 用于模拟机器学习模型

# 安全性
cryptography==42.0.5   # 用于模拟加密操作

# 开发工具
flake8==7.0.0          # 用于代码风格检查
pytest==8.1.1          # 用于模拟单元测试